fix: 사용자 null일때 에러 포착을 위해 컴포넌트 따로 생성
